+++ R/geom-label-npc.r
@@ -57,13 +57,14 @@ GeomLabelNpc <- ggplot2::ggproto("GeomLabelNpc", ggplot2::Geom,
 
   default_aes = ggplot2::aes(
     colour = "black", fill = "white", size = 3.88, angle = 0, hjust = "inward",
-    vjust = "inward", alpha = NA, family = "", fontface = 1, lineheight = 1.2
+    vjust = "inward", alpha = NA, family = "", fontface = 1, lineheight = 1.2,
+    linetype = "solid"
   ),
 
-  draw_panel = function(data, panel_params, coord, parse = FALSE,
+  draw_panel = function(self, data, panel_params, coord, parse = FALSE,
                         na.rm = FALSE,
-                        label.padding = grid::unit(0.25, "lines"),
-                        label.r = grid::unit(0.15, "lines"),
+                        label.padding = unit(0.25, "lines"),
+                        label.r = unit(0.15, "lines"),
                         label.size = 0.25,
                         size.unit = "mm") {
 
@@ -75,19 +76,55 @@ GeomLabelNpc <- ggplot2::ggproto("GeomLabelNpc", ggplot2::Geom,
     data$x <- ranges$x[1] + data$npcx * (ranges$x[2] - ranges$x[1])
     data$y <- ranges$y[1] + data$npcy * (ranges$y[2] - ranges$y[1])
 
-    ggplot2::GeomLabel$draw_panel(data = data,
-                                  panel_params = panel_params,
-                                  coord = coord,
-                                  parse = parse,
-                                  na.rm = na.rm,
-                                  label.padding = label.padding,
-                                  label.r = label.r,
-                                  label.size = label.size)
-    # for ggplot2 >= 3.5.0 add conditionally:
-    # add size.unit = size.unit
+    lab <- data$label
+
+    if (parse) {
+      lab <- parse_safe(as.character(lab))
+    }
+
+    data <- coord$transform(data, panel_params)
+    if (is.character(data$vjust)) {
+      data$vjust <- compute_just(data$vjust, data$y)
+    }
+    if (is.character(data$hjust)) {
+      data$hjust <- compute_just(data$hjust, data$x)
+    }
+    if (!inherits(label.padding, "margin")) {
+      label.padding <- rep(label.padding, length.out = 4)
+    }
+
+    size.unit <- resolve_text_unit(size.unit)
+
+    grobs <- lapply(1:nrow(data), function(i) {
+      row <- data[i, , drop = FALSE]
+      labelGrob(lab[i],
+                x = grid::unit(row$x, "native"),
+                y = grid::unit(row$y, "native"),
+                just = c(row$hjust, row$vjust),
+                padding = label.padding,
+                r = label.r,
+                angle = row$angle,
+                text.gp = grid::gpar(
+                  col = row$colour,
+                  fontsize = row$size * size.unit,
+                  fontfamily = row$family,
+                  fontface = row$fontface,
+                  lineheight = row$lineheight
+                ),
+                rect.gp = grid::gpar(
+                  col = if (isTRUE(all.equal(label.size, 0))) NA else row$colour,
+                  fill = ggplot2::alpha(row$fill, row$alpha),
+                  lwd = label.size * .pt,
+                  lty = row$linetype
+                )
+      )
+    })
+    class(grobs) <- "gList"
+
+    ggname("geom_label", grobTree(children = grobs))
   },
 
   draw_key =  function(...) {
-    grid::nullGrob()
+    grid::nullGrob()  # geom meant to be used for annotations
   }
 )